src/models/requirements.txt
argparse
pandas
Pillow
matplotlib
numpy
scikit-learn
torch
torchvision
tqdm
google-cloud-storage
